pyproject.toml
[project]
name = "paperbanana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai",
    "streamlit",
    "asyncio",
    "aiofiles",
    "pillow",
    "numpy",
    "tqdm",
    "json-repair",
    "anthropic",
    "openai",
    "matplotlib",
    "python-dotenv",
    "pyyaml",
    "google-auth",
    "huggingface-hub>=1.6.0",
]
